ldone with the front end test

+++ src/main/java/Parcel.java
@@ -6,7 +6,7 @@ public class Parcel {
   private int mWeight;
   private int mVolume;
   private int mMiles;
-  private double mShipping;
+  private int mShipping;
 
   public Parcel(int length, int width, int height, int weight, int miles) {
     mLength = length;
@@ -32,8 +32,8 @@ public class Parcel {
   public int getVolume() {
    return mVolume;
   }
-  public double getShipping() {
-    mShipping = (mVolume * 0.10 * mWeight * mMiles);
+  public int getShipping() {
+    mShipping = (int)(mVolume * 0.10 * mWeight * mMiles);
     return mShipping;
   }
 }
